put tree model in its own folder
